requirements.txt
Flask
torch
torchvision
open_clip_torch
pandas
Pillow
numpy